6. Size is not grandeur, and territory does not make a nation

+++ 0x00-hello_world/6-size.c
@@ -1,14 +1,16 @@
 #include <stdio.h>
+
 /**
- * main - program printing the size of various types based on computer run on
+ * main - program printing size based on various systems
  * Return: always 0
  */
+
 int main(void)
 {
 printf("Size of a char: %zu byte(s)\n", (unsigned long)sizeof(char));
 printf("Size of an int: %zu byte(s)\n", (unsigned long)sizeof(int));
 printf("Size of a long int: %zu byte(s)\n", (unsigned long)sizeof(long int));
-printf("Size of a long long int: %zu byte(s)\n", (unsigned long)sizeof(long long int));
+printf("Size of a long long int: %zu byte(s)\n", sizeof(long long int));
 printf("Size of a float: %zu byte(s)\n", (unsigned long)sizeof(float));
 return (0);
 }
